Início do adicionar perfil e menus extras ao cadastro de profissional.

+++ sishosp/src/br/gov/al/maceio/sishosp/acl/dao/MenuDAO.java
@@ -763,7 +763,7 @@ public class MenuDAO {
 		return lista;
 	}
 
-	public ArrayList<Menu> listarMenuItemSourcerUser(Integer idPerfil)
+	public ArrayList<Menu> listarMenuItemSourcerUser(Long idPerfil)
 			throws ProjetoException {
 
 		String sql = "select me.id, me.descricao, me.codigo, me.indice, me.tipo, "
@@ -788,7 +788,7 @@ public class MenuDAO {
 		try {
 			conexao = ConnectionFactory.getConnection();
 			PreparedStatement stmt = conexao.prepareStatement(sql);
-			stmt.setInt(1, idPerfil);
+			stmt.setLong(1, idPerfil);
 			ResultSet rs = stmt.executeQuery();
 
 			while (rs.next()) {
